changed credit data type to double; allowed decimal type keyboard for credit text field

+++ lib/screens/club_screen/member_credit_history.dart
@@ -349,7 +349,7 @@ class _ClubActivityCreditScreenState extends State<ClubActivityCreditScreen> {
                                     context: context,
                                     clubCode: widget.clubCode,
                                     playerUuid: widget.playerId,
-                                    credits: member.availableCredit.toInt());
+                                    credits: member.availableCredit.toDouble());
                                 if (ret) {
                                   changed = true;
                                   fetchData();
